+++ js/src/contract_gen.ts
@@ -34,6 +34,11 @@ export const contract = {
             "1080p"
           ]
         },
+        "seed": {
+          "min": -1,
+          "max": 2147483647,
+          "type": "integer"
+        },
         "source_image_urls": {
           "max_items": 2
         }
@@ -216,6 +221,11 @@ export const contract = {
             "720p",
             "1080p"
           ]
+        },
+        "seed": {
+          "min": -1,
+          "max": 2147483647,
+          "type": "integer"
         }
       }
     }